add to cart + dahsboard (all endpoints are done)
