+++ restapi/controllers/__init__.py
@@ -4,4 +4,5 @@ from . import controllers
 from . import get
 from . import customers
 from . import sales
+from . import sales_reporting
 from . import security
